requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic>=2.0,<3

# Course chatbot v0–v3 (LangChain + Flask UI); see chatbot_v0.py
flask>=3.0.0,<4
langchain-core>=0.3.0,<1
langchain-openai>=0.2.0,<1
python-dotenv>=1.0.0,<2
